modulargame/x/modulargame/types/message_create_fellow_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"modulargame/testutil/sample"
)

func TestMsgCreateFellow_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateFellow
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateFellow{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateFellow{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
